app/core/dependency.py
from typing import Any

import jwt
import orjson
from fastapi import Depends, Request
from fastapi.security import OAuth2PasswordBearer

from app.core.code import Code
from app.core.ctx import CTX_USER_ID, CTX_X_REQUEST_ID
from app.core.exceptions import (
    HTTPException,
)
from app.log import log
from app.models.system import User, StatusType
from app.settings import APP_SETTINGS
from app.utils.tools import check_url

oauth2_schema = OAuth2PasswordBearer(tokenUrl="/auth/token", auto_error=False)


def check_token(token: str) -> tuple[bool, str, Any]:
    try:
        options = {"verify_signature": True, "verify_aud": False, "exp": True}
        decode_data = jwt.decode(token, APP_SETTINGS.SECRET_KEY, algorithms=[APP_SETTINGS.JWT_ALGORITHM], options=options)
        return True, Code.SUCCESS, decode_data
    except jwt.DecodeError:
        return False, Code.INVALID_TOKEN, "无效的Token"
    except jwt.ExpiredSignatureError:
        return False, Code.TOKEN_EXPIRED, "登录已过期"
    except Exception as e:
        return False, Code.INVALID_TOKEN, f"{repr(e)}"


class AuthControl:
    @classmethod
    async def is_authed(cls, token: str = Depends(oauth2_schema)) -> User | None:
        if not token:
            raise HTTPException(code=Code.INVALID_TOKEN, msg="Authentication failed, token does not exists in the request.")
        user_id = CTX_USER_ID.get()
        if user_id == 0:
            status, code, decode_data = check_token(token)
            if not status:
                raise HTTPException(code=code, msg=decode_data)

            if decode_data["data"]["tokenType"] != "accessToken":
                raise HTTPException(code=Code.INVALID_SESSION, msg="The token is not an access token")

            user_id = decode_data["data"]["userId"]

        user = await User.filter(id=user_id).first()
        if not user:
            raise HTTPException(code=Code.INVALID_SESSION, msg=f"Authentication failed, the user_id: {user_id} does not exists in the system.")
        CTX_USER_ID.set(int(user_id))
        return user


class PermissionControl:
    @classmethod
    async def has_permission(cls, request: Request, current_user: User = Depends(AuthControl.is_authed)) -> None:
        redis = request.app.state.redis
        cache_key = f"user_perms:{current_user.id}"
        
        # Try to get from cache
        try:
            cached_perms = await redis.get(cache_key)
            if cached_perms:
                permission_apis = orjson.loads(cached_perms)
            else:
                permission_apis = None
        except Exception:
            permission_apis = None

        if permission_apis is None:
            await current_user.fetch_related("by_user_roles")
            user_roles_codes: list[str] = [r.role_code for r in current_user.by_user_roles]
            if "R_SUPER" in user_roles_codes:  # 超级管理员
                # Cache super status or just return
                return

            if not current_user.by_user_roles:
                raise HTTPException(code=Code.PERMISSION_DENIED, msg="The user is not bound to a role")

            apis = [await role.by_role_apis for role in current_user.by_user_roles]
            # permission_apis = list(set((api.api_method.value, api.api_path, api.status_type) for api in sum(apis, [])))
            # orjson can't serialize sets/enums directly without help, let's normalize to strings/dicts
            permission_apis = []
            for api in sum(apis, []):
                permission_apis.append({
                    "method": api.api_method.value.lower(),
                    "path": api.api_path,
                    "status": api.status_type.value
                })
            
            # Cache for 1 hour (3600 seconds)
            try:
                await redis.setex(cache_key, 3600, orjson.dumps(permission_apis))
            except Exception:
                pass

        # Check permission
        method = request.method.lower()
        path = request.url.path
        
        # Check against cached or freshly fetched perms
        # Special case for super admin (handled above already if freshly fetched, 
        # but if from cache we should have a flag)
        
        for api in permission_apis:
            if api["method"] == method and check_url(api["path"], path):
                if api["status"] == StatusType.disable.value:
                    raise HTTPException(code=Code.API_DISABLED, msg=f"The API has been disabled, method: {method} path: {path}")
                return

        log.error("*" * 20)
        log.error(f"Permission denied, method: {method.upper()} path: {path}")
        log.error(f"x-request-id: {CTX_X_REQUEST_ID.get()}")
        log.error("*" * 20)
        raise HTTPException(code=Code.PERMISSION_DENIED, msg=f"Permission denied, method: {method} path: {path}")



DependAuth = Depends(AuthControl.is_authed)
DependPermission = Depends(PermissionControl.has_permission)
