pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project]
name = "fast-soy-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.2",
    "fastapi-cache2[redis]>=0.2.2",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "orjson>=3.11.7",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "starlette==0.52.1",
    "tortoise-orm>=0.25.4",
    "uvicorn>=0.42.0",
    "aiomysql>=0.3.2",
]

[tool.aerich]
tortoise_orm = "app.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.pyright]
include = ["app"]
exclude = [
    "web",
    ".venv",
    "**/.*",
    "**/__pycache__",
    "**/node_modules",
    "src/experimental",
    "src/typestubs",
]
pythonVersion = "3.12"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
  { root = "./tests/python_3_12", pythonVersion = "3.12", extraPaths = [
    "./",
  ] },
  { root = "./tests", extraPaths = [
    "./",
  ] },
  { root = "./" },
]

typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true
